[IMP] Moved endpoint_removed out of util.py into blueprints namespace
